requirements.txt
requests>=2.28.0
pandas>=1.5.0
pyarrow>=10.0.0
pytest>=7.0.0
numpy>=1.23.0
python-dotenv>=1.0.0
